requirements.txt
Babel==0.9.6
Beaker==1.6.4
Chameleon==2.11
FormEncode==1.2.6
Mako==0.8.0
MarkupSafe==0.15
PasteDeploy==1.5.0
Pillow==2.0.0
Pygments==1.6
SQLAlchemy==0.8.2
Unidecode==0.04.12
WebOb==1.2.3
alembic==0.5.0
colander==1.0a2
deform==0.9.7
deform-bootstrap==0.2.9
docopt==0.6.1
fanstatic==1.0a5
html2text==3.200.3
js.angular==1.1.2
js.bootstrap==2.3.1
js.chosen==0.9.11
js.deform==0.9.7
js.deform-bootstrap==0.2.6
js.fineuploader==3.3.0
js.html5shiv==3.6.2-1
js.jquery==1.9.1
js.jquery-form==3.09
js.jquery-maskedinput==1.3.1
js.jquery-maskmoney==1.4.1
js.jquery-tablednd==0.4
js.jquery-timepicker-addon==1.3-1
js.jqueryui==1.10.3
js.jqueryui-tagit==2.0.24-2
js.tinymce==4.0.2
kotti-tinymce==0.4.1
kotti_blog==0.4
kotti_media==0.5.1
lingua==1.5
newrelic==2.18.1.15
peppercorn==0.4
plone.scale==1.3.1
polib==1.0.3
psycopg2==2.5.2
py-bcrypt==0.3
pyramid==1.4.3
pyramid-beaker==0.7
pyramid-debugtoolbar==1.0.6
pyramid-deform==0.2a5
pyramid-mailer==0.11
pyramid-tm==0.7
pyramid-zcml==1.0.0
pytz==2013b
repoze.lru==0.6
repoze.sendmail==4.0
repoze.workflow==0.6.1
repoze.zcml==0.4
transaction==1.4.1
translationstring==1.1
venusian==1.0a8
waitress==0.8.5
xlrd==0.9.2
xlwt==0.7.5
zope.component==4.1.0
zope.configuration==4.0.2
zope.deprecation==4.0.2
zope.event==4.0.2
zope.i18nmessageid==4.0.2
zope.interface==4.0.5
zope.schema==4.3.2
zope.sqlalchemy==0.7.2
